add dynamic SUT support to more modelgauge commands
